split the tests for different models
