+++ .gitignore
@@ -163,5 +163,9 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 
+# Example data
+datasets/
+temp/
+
 # other
 .DS_Store
